+++ src/components/common/modal/ShareModal.styled.ts
