+++ src/components/common/modal/ShareModal.tsx
@@ -1,4 +1,4 @@
-import * as Styled from "./styled";
+import * as Styled from "./ShareModal.styled";
 import { GreenBtn } from "../button/GreenBtn";
 
 interface ModalProps {
